+++ src/ts/utils/datastructures/Color.ts
@@ -14,7 +14,9 @@
  * @modified 2021-11-07 Did the same with `saturate` and `desaturate`.
  * @modified 2021-11-07 Did the same with the `fadein` and `fadeout` functions.
  * @modified 2021-11-07 Added setRed, setGreen, setBlue, setHue, setSaturation, setLiminance functions.
- * @version 0.0.10
+ * @modified 2022-05-11 Modified the `clone` function by just copying the numeric calues, not re-calculating the whole color.
+ * @modified 2022-05-11 Fixed the `interpolate` function.
+ * @version 0.0.11
  **/
 
 /**
@@ -698,7 +700,16 @@ export class Color {
    * @return {Color} A clone of this color (in RGB mode).
    */
   clone(): Color {
-    return Color.makeRGB(this.r, this.g, this.b, this.a);
+    // return Color.makeRGB(this.r, this.g, this.b, this.a);
+    const col = new Color();
+    col.r = this.r;
+    col.g = this.g;
+    col.b = this.b;
+    col.a = this.a;
+    col.h = this.h;
+    col.s = this.s;
+    col.l = this.l;
+    return col;
   }
 
   /**
@@ -712,10 +723,10 @@ export class Color {
    * @return {Color} A clone of this color (in RGB mode).
    */
   interpolate(c: Color, t: number): Color {
-    this.r += (c.r - c.r) * t;
-    this.g += (c.g - c.g) * t;
-    this.b += (c.b - c.b) * t;
-    this.a += (c.a - c.a) * t;
+    this.r += (c.r - this.r) * t;
+    this.g += (c.g - this.g) * t;
+    this.b += (c.b - this.b) * t;
+    this.a += (c.a - this.a) * t;
     return this;
   }
 } // END class